Add UserDelegate to allow riding on top of the push/pull mechanism
memberlist occasionally does full push/pull syncs over TCP. This
can be ridden using the Delegate in the memberlist config. However,
serf overrides this with its own Delegate to manage cluster memberships.

This commit adds a UserDelegate to the serf config. This is called
from the serf Delegate, allowing the user to ride on top of the
push/pull mechanism.

Signed-off-by: sinkingpoint <colin@quirl.co.nz>